chore(fs): Refactor firestore error logic
