+++ src/firestore/index.ts
@@ -221,4 +221,4 @@ export function initializeFirestore(
   return firestoreService.initializeDatabase(databaseId, settings);
 }
 
-export { FirebaseFirestoreError } from '../utils/error';
+export { FirebaseFirestoreError } from './error';
